Extracted controllers into separate files for requests, spaces and users.
