+++ app/controllers/users.rb
@@ -1,23 +1,5 @@
 class MakersBnb < Sinatra::Base
 
-
-get '/edit/:space_id' do
-  unless current_user
-    redirect '/sessions/new'
-  end
-  @space = Space.get(params[:space_id])
-  if current_user.id == @space.user_id
-    erb :'spaces/edit'
-  else
-    redirect '/users/account'
-  end
-end
-
-post '/edit'do
-Space.get(params[:space_id]).update(name: params[:name], description: params[:description], price: params[:price])
-redirect to 'users/account'
-end
-
 get '/users/new' do
   @user = User.new
   erb :'users/new'
@@ -68,111 +50,6 @@ delete '/sessions' do
   flash[:notice] = "See you later, #{current_user.name}"
   session[:user_id] = nil
   redirect '/'
-end
-
-get '/spaces' do
-  if session[:space_array]
-    @spaces = session[:space_array].map do |space_id|
-      Space.get(space_id)
-    end
-  else
-    @spaces = Space.all
-  end
-  session[:space_array] = nil
-  erb :'spaces/index'
-end
-
-get '/requests/:space' do
-  @space = Space.get(params[:space])
-  available_dates = AvailableDate.all(space_id: params[:space])
-  @formatted_dates = []
-  available_dates.each do |date|
-    @formatted_dates << date.available_date.strftime
-  end
-  session[:available_dates] = @formatted_dates
-  session[:array] = @formatted_dates.to_json
-  erb :'requests/new'
-end
-
-post '/requests/new' do
-
-  requested_date = Date.parse(params[:date_requested]) rescue nil
-
-  unless requested_date
-    flash[:notice] = 'Please enter a valid date'
-    redirect '/spaces'
-  end
-
-  unless session[:available_dates].include?(Date.parse(params[:date_requested]).strftime)
-    flash[:notice] = 'Space not available on the requested date'
-    redirect '/spaces'
-  end
-
-  if current_user
-    unless current_user.id != Space.get(params[:space_id]).user_id
-      flash[:notice] = 'This is your own space'
-      redirect '/spaces'
-    end
-    date = Date.parse(params[:date_requested])
-    space = Space.get(params[:space_id])
-    request = Request.create(user_id: current_user.id, space_id: params[:space_id], date_requested: date, owner_id: space.user_id, space_name: space.name, requested_by: current_user.name)
-    flash.keep[:notice] = 'Booking requested'
-    redirect '/users/account'
-  else
-    flash[:notice] = 'Please log in to request a space'
-    redirect '/spaces'
-  end
-end
-
-get '/requests/confirm/:request_id' do
-
-end
-
-get '/requests' do
-  if current_user
-    @users_requests = Request.all(user_id: current_user.id)
-    @received_requests = Request.all(owner_id: current_user.id)
-    @request_display = @users_requests.map do |x|
-      Space.get(x.space_id)
-    end
-    erb :'requests/index'
-  else
-    redirect to '/sessions/new'
-  end
-end
-
-post '/spaces' do
-  validate_dates(params[:available_from], params[:available_to], '/spaces')
-  find_available_spaces(params[:available_from], params[:available_to])
-  redirect to '/spaces'
-end
-
-get '/spaces/new' do
-  if current_user
-    @todays_date = Date.today
-    erb :'spaces/new'
-  else
-    flash.keep[:errors] = ['Plese login to list a space']
-    redirect '/sessions/new'
-  end
-end
-
-post '/spaces/new' do
-
-  validate_dates(params[:available_from],params[:available_to], 'spaces/new')
-  @space = Space.create(name: params[:name], description: params[:description], price: params[:price], available_from: params[:available_from], available_to: params[:available_to], user_id: current_user.id)
-
-  if @space.save
-    date_from = @space.available_from
-    date_to = @space.available_to
-    (date_from..date_to).each do |date|
-      AvailableDate.create(available_date: date, space_id: @space.id)
-    end
-    redirect to '/spaces'
-  else
-    flash[:notice] = 'Please complete the required fields'
-    redirect to '/spaces/new'
-  end
 end
 
 end
